# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Changqingjing API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/admin/scenics/{entryId}/draft": {
      "put": {
        "tags": [
          "admin-scenic-controller"
        ],
        "operationId": "saveDraft",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveScenicDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminScenicContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/contents/home-videos/{entryId}/draft": {
      "put": {
        "tags": [
          "admin-home-video-content-controller"
        ],
        "operationId": "saveDraft_1",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveHomeVideoDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeVideoContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/home-hero/draft": {
      "put": {
        "tags": [
          "admin-home-hero-content-controller"
        ],
        "operationId": "saveDraft_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveHomeHeroDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeHeroContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/contents/company/draft": {
      "put": {
        "tags": [
          "admin-company-content-controller"
        ],
        "operationId": "saveDraft_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCompanyDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminCompanyContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/scenics/{entryId}/unpublish": {
      "post": {
        "tags": [
          "admin-scenic-controller"
        ],
        "operationId": "unpublish",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminScenicContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/scenics/{entryId}/publish": {
      "post": {
        "tags": [
          "admin-scenic-controller"
        ],
        "operationId": "publish",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminScenicContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/products/{productId}/unpublish": {
      "post": {
        "tags": [
          "admin-product-controller"
        ],
        "operationId": "unpublish_1",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminProductContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{productId}/publish": {
      "post": {
        "tags": [
          "admin-product-controller"
        ],
        "operationId": "publish_1",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminProductContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/product-categories/{categoryId}/unpublish": {
      "post": {
        "tags": [
          "admin-product-category-controller"
        ],
        "operationId": "unpublish_2",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminProductCategoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/product-categories/{categoryId}/publish": {
      "post": {
        "tags": [
          "admin-product-category-controller"
        ],
        "operationId": "publish_2",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminProductCategoryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/contents/home-videos/{entryId}/unpublish": {
      "post": {
        "tags": [
          "admin-home-video-content-controller"
        ],
        "operationId": "unpublish_3",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeVideoContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/home-videos/{entryId}/publish": {
      "post": {
        "tags": [
          "admin-home-video-content-controller"
        ],
        "operationId": "publish_3",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeVideoContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/home-hero/unpublish": {
      "post": {
        "tags": [
          "admin-home-hero-content-controller"
        ],
        "operationId": "unpublish_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeHeroContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/home-hero/publish": {
      "post": {
        "tags": [
          "admin-home-hero-content-controller"
        ],
        "operationId": "publish_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeHeroContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/cooperation/unpublish": {
      "post": {
        "tags": [
          "admin-cooperation-controller"
        ],
        "operationId": "unpublish_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminCooperationContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/cooperation/publish": {
      "post": {
        "tags": [
          "admin-cooperation-controller"
        ],
        "operationId": "publish_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminCooperationContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/company/unpublish": {
      "post": {
        "tags": [
          "admin-company-content-controller"
        ],
        "operationId": "unpublish_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminCompanyContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/contents/company/publish": {
      "post": {
        "tags": [
          "admin-company-content-controller"
        ],
        "operationId": "publish_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminCompanyContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/scenics/{entryId}/preview": {
      "get": {
        "tags": [
          "admin-scenic-controller"
        ],
        "operationId": "preview",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminScenicRevisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/products/{productId}/preview": {
      "get": {
        "tags": [
          "admin-product-controller"
        ],
        "operationId": "preview_1",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminProductRevisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/contents/home-videos/{entryId}/preview": {
      "get": {
        "tags": [
          "admin-home-video-content-controller"
        ],
        "operationId": "preview_2",
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAdminHomeVideoRevisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SaveScenicDraftRequest": {
        "required": [
          "openStatus",
          "summary",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "summary": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "coverMediaId": {
            "type": "string",
            "format": "uuid"
          },
          "blocks": {
            "maxItems": 100,
            "minItems": 0,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenicContentBlock"
            }
          },
          "openStatus": {
            "type": "string",
            "enum": [
              "OPEN",
              "PAUSED"
            ]
          },
          "displayOrder": {
            "maximum": 10000,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "displayName": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "locationSelectionId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedVersion": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ScenicContentBlock": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HEADING",
              "PARAGRAPH",
              "IMAGE"
            ]
          },
          "text": {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string"
          },
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "altText": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "AdminScenicContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "draft": {
            "$ref": "#/components/schemas/AdminScenicRevisionResponse"
          },
          "published": {
            "$ref": "#/components/schemas/AdminScenicRevisionResponse"
          }
        }
      },
      "AdminScenicRevisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revisionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "coverMediaId": {
            "type": "string",
            "format": "uuid"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenicContentBlock"
            }
          },
          "openStatus": {
            "type": "string",
            "enum": [
              "OPEN",
              "PAUSED"
            ]
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "$ref": "#/components/schemas/ScenicLocation"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseAdminScenicContentResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminScenicContentResponse"
          }
        }
      },
...
      "CompanyContentBlock": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HEADING",
              "PARAGRAPH",
              "IMAGE"
            ]
          },
          "text": {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string"
          },
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "altText": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "AdminProductContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "draft": {
            "$ref": "#/components/schemas/AdminProductRevisionResponse"
          },
          "published": {
            "$ref": "#/components/schemas/AdminProductRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminProductContentResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminProductContentResponse"
          }
        }
      },
...
      "AdminProductCategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "hasUnpublishedChanges": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseAdminProductCategoryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminProductCategoryResponse"
          }
        }
      },
      "SaveHomeVideoDraftRequest": {
        "required": [
          "coverMediaId",
          "title",
          "videoMediaId"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "videoMediaId": {
            "type": "string",
            "format": "uuid"
          },
          "coverMediaId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedVersion": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AdminHomeVideoContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "draft": {
            "$ref": "#/components/schemas/AdminHomeVideoRevisionResponse"
          },
          "published": {
            "$ref": "#/components/schemas/AdminHomeVideoRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminHomeVideoContentResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminHomeVideoContentResponse"
          }
        }
      },
      "SaveHomeHeroDraftRequest": {
        "required": [
          "coverMediaId"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "subtitle": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "coverMediaId": {
            "type": "string",
            "format": "uuid"
          },
          "focusX": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "focusY": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "expectedVersion": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AdminHomeHeroContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "draft": {
            "$ref": "#/components/schemas/AdminHomeHeroRevisionResponse"
          },
          "published": {
            "$ref": "#/components/schemas/AdminHomeHeroRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminHomeHeroContentResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminHomeHeroContentResponse"
          }
        }
      },
...
      "AdminCooperationContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "draft": {
            "$ref": "#/components/schemas/AdminCooperationRevisionResponse"
          },
          "published": {
            "$ref": "#/components/schemas/AdminCooperationRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminCooperationContentResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminCooperationContentResponse"
          }
        }
      },
      "SaveCompanyDraftRequest": {
        "required": [
          "blocks",
          "summary",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "summary": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "coverMediaId": {
            "type": "string",
            "format": "uuid"
          },
          "blocks": {
            "maxItems": 100,
            "minItems": 0,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyContentBlock"
            }
          },
          "expectedVersion": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AdminCompanyContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string"
          },
          "firstPublishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "draft": {
            "$ref": "#/components/schemas/AdminCompanyRevisionResponse"
          },
          "published": {
            "$ref": "#/components/schemas/AdminCompanyRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminCompanyContentResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminCompanyContentResponse"
          }
        }
      },
...
      "ContentVersionRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ApiResponseAdminScenicRevisionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminScenicRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminProductRevisionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminProductRevisionResponse"
          }
        }
      },
...
      "ApiResponseAdminHomeVideoRevisionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminHomeVideoRevisionResponse"
          }
        }
      },
...
    }
  }
}